+++ Source/DotNetWorkQueue.Transport.SQLite.Integration.Tests/ConsumerAsync/ConsumerAsyncExpiredMessage.cs
@@ -1,0 +1,33 @@
+using DotNetWorkQueue.Configuration;
+using DotNetWorkQueue.IntegrationTests.Shared;
+using DotNetWorkQueue.IntegrationTests.Shared.Producer;
+using DotNetWorkQueue.Queue;
+using DotNetWorkQueue.Transport.SQLite.Basic;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace DotNetWorkQueue.Transport.SQLite.Integration.Tests.ConsumerAsync
+{
+    [TestClass]
+    public class ConsumerAsyncExpiredMessage
+    {
+        [TestMethod]
+        [DataRow(100, 60, 5, 2, true, false),
+         DataRow(100, 60, 5, 2, false, false),
+         DataRow(10, 60, 5, 2, false, true)]
+        public void Run(int messageCount, int timeOut, int workerCount, int readerCount, bool inMemoryDb, bool enableChaos)
+        {
+            using (var connectionInfo = new IntegrationConnectionInfo(inMemoryDb))
+            {
+                var queueName = GenerateQueueName.Create();
+                var consumer =
+                    new DotNetWorkQueue.IntegrationTests.Shared.ConsumerAsync.Implementation.
+                        ConsumerAsyncExpiredMessage();
+                consumer.Run<SqLiteMessageQueueInit, FakeMessage, SqLiteMessageQueueCreation>(new QueueConnection(queueName, connectionInfo.ConnectionString),
+                    messageCount, timeOut, workerCount, readerCount, enableChaos, x => Helpers.SetOptions(x,
+                        true, true, true,
+                        false, true, true, false),
+                    Helpers.GenerateData, Helpers.Verify, Helpers.VerifyQueueCount);
+            }
+        }
+    }
+}