+++ Source/DotNetWorkQueue.Transport.SqlServer.IntegrationTests/ConsumerAsync/ConsumerAsyncExpiredMessage.cs
@@ -1,0 +1,29 @@
+using DotNetWorkQueue.Configuration;
+using DotNetWorkQueue.IntegrationTests.Shared;
+using DotNetWorkQueue.IntegrationTests.Shared.Producer;
+using DotNetWorkQueue.Transport.SqlServer.Basic;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace DotNetWorkQueue.Transport.SqlServer.IntegrationTests.ConsumerAsync
+{
+    [TestClass]
+    public class ConsumerAsyncExpiredMessage
+    {
+        [TestMethod]
+        [DataRow(100, 60, 5, 2, false, false),
+        DataRow(100, 60, 5, 2, true, false),
+        DataRow(10, 60, 5, 2, true, true)]
+        public void Run(int messageCount, int timeOut, int workerCount, int readerCount, bool useTransactions, bool enableChaos)
+        {
+            var queueName = GenerateQueueName.Create();
+            var consumer =
+                new DotNetWorkQueue.IntegrationTests.Shared.ConsumerAsync.Implementation.ConsumerAsyncExpiredMessage();
+            consumer.Run<SqlServerMessageQueueInit, FakeMessage, SqlServerMessageQueueCreation>(new QueueConnection(queueName, ConnectionInfo.ConnectionString),
+                messageCount, timeOut, workerCount, readerCount, enableChaos, x => Helpers.SetOptions(x,
+                    true, !useTransactions, useTransactions,
+                    true,
+                    false, !useTransactions, true, false),
+                Helpers.GenerateData, Helpers.Verify, Helpers.VerifyQueueCount);
+        }
+    }
+}